Increased EV charger turn-off window
If the EV charger turns off at the very end of the 30-minute billing period, the measured power to the car may take a little time to reduce, so the battery charges for a further 30 minutes at the peak rate. Changed to sampling every minute for 0, 1 and 2 minutes past the period end.